% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Bar_Charts.R
\name{PM_Bar_Charts}
\alias{PM_Bar_Charts}
\title{Creates Poshmark Bar Chart}
\usage{
PM_Bar_Charts(chart_data, factor_var, x_axis, chart_title, chart_type)
}
\arguments{
\item{chart_data}{The Chart Data.}

\item{factor_var}{The Factor Variable.}

\item{x_axis}{The X-Axis.}

\item{chart_title}{The Chart Title.}

\item{chart_type}{The Chart Type}
}
\value{
Chart for presentation.
}
\description{
This function Creates Poshmark Bar Chart.
}
\details{
Requires: ggplot2, scales
}
