% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_DT_Summary_Table.R
\name{PM_DT_Summary_Table}
\alias{PM_DT_Summary_Table}
\title{Creates the Poshmark Summary Data Table}
\usage{
PM_DT_Summary_Table(summary_table, table_type)
}
\arguments{
\item{summary_table}{The table to be used for the report.}

\item{table_type}{The table type for the report.}
}
\value{
A Poshmark Summary Data Table.
}
\description{
This function creates a Poshmark Summary Data Table.
}
\details{
Requires: DT
}
\examples{

}
