% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Table_Delete_Record.R
\name{PM_Sales_Table_Delete_Record}
\alias{PM_Sales_Table_Delete_Record}
\title{Deletes Poshmark Sales Activity Record from Database}
\usage{
PM_Sales_Table_Delete_Record(inv_record)
}
\arguments{
\item{inv_record}{}
}
\value{

}
\description{
This function Deletes Poshmark Sales Activity Record from Database.
}
\details{
Requires: RPostgres
}
