% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Table_Sort.R
\name{PM_Sales_Table_Sort}
\alias{PM_Sales_Table_Sort}
\title{Sorts the Poshmark Sales Table}
\usage{
PM_Sales_Table_Sort(pm_sales)
}
\arguments{
\item{inv_table}{Inventory Table to Sort}
}
\value{
inv_table Sorted Inventory Table
}
\description{
This function Sorts the Poshmark Sales Table.
}
