% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Upload_Format_Summary.R
\name{PM_Sales_Upload_Format_Summary}
\alias{PM_Sales_Upload_Format_Summary}
\title{Format Poshmark Upload Panel}
\usage{
PM_Sales_Upload_Format_Summary(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity Table}
}
\value{
alert_text HTML Formatted Panel
}
\description{
This function formats the Poshmark Upload Panel.
}
\details{
Requires: shiny
}
\examples{

}
