% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Upload_Item_Cost.R
\name{PM_Sales_Upload_Item_Cost}
\alias{PM_Sales_Upload_Item_Cost}
\title{Get Item Cost for Poshmark Sales Activity}
\usage{
PM_Sales_Upload_Item_Cost(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity Report}
}
\value{
sales_activity Poshmark Sales Activity Report
}
\description{
This function finds the item cost for the Poshmark Sales Activity table.
}
\details{
Requires:
}
\examples{

}
