% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Upload_Process.R
\name{PM_Sales_Upload_Process}
\alias{PM_Sales_Upload_Process}
\title{Process Poshmark Sales Activity File}
\usage{
PM_Sales_Upload_Process(t_datapath)
}
\arguments{
\item{t_datapath}{The file to be read}
}
\value{
sales_activity The Poshmark Sales Activity table as data frame.
}
\description{
This function Processes the Poshmark Sales Activity File.
}
\details{
Requires: utils, dplyr, lubridate
}
\examples{

}
