% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Upload_Summary_Stats.R
\name{PM_Sales_Upload_Summary_Stats}
\alias{PM_Sales_Upload_Summary_Stats}
\title{Creates Poshmark Summary Stats for File Upload}
\usage{
PM_Sales_Upload_Summary_Stats(sales_activity)
}
\arguments{
\item{sales_activity}{Poshmark Sales Activity table.}
}
\value{
summary_stats Poshmark summary statistics.
}
\description{
This function Creates Poshmark Summary Stats for File Upload.
}
\details{
Requires: dplyr
}
\examples{

}
