% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Upload_Update_Inventory.R
\name{PM_Sales_Upload_Update_Inventory}
\alias{PM_Sales_Upload_Update_Inventory}
\title{Update Inventory Table with Poshmark Sales Activity Table}
\usage{
PM_Sales_Upload_Update_Inventory(d_sales_activity)
}
\arguments{
\item{d_sales_activity}{Poshmark Sales Activity Report}
}
\value{
Updates Inventory Table
}
\description{
This function Updates the Inventory Table with the Poshmark Sales Activity Table.
}
\details{
Requires:
}
