% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Sales_Upload_Validate.R
\name{PM_Sales_Upload_Validate}
\alias{PM_Sales_Upload_Validate}
\title{Check the Validity of the Input Poshmark Sales Activity File}
\usage{
PM_Sales_Upload_Validate(t_datapath)
}
\arguments{
\item{t_datapath}{The file to be read.}
}
\value{
valid_report, valid_user Logical values in array.
}
\description{
This function Checks the Validity of the Input Poshmark Sales Activity File.
}
\examples{

}
