% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Table_Monthly_Activity.R
\name{PM_Table_Monthly_Activity}
\alias{PM_Table_Monthly_Activity}
\title{Creates a Poshmark Monthly Activity Table}
\usage{
PM_Table_Monthly_Activity(pm_sales, flt_year, flt_month)
}
\arguments{
\item{pm_sales}{The Poshmark Sales Table for the user.}
}
\value{
A Poshmark Monthly Activity Table as a data frame.
}
\description{
This function creates a Poshmark Monthly Activity Table from the Inventory Table.
}
\details{
Requires: dplyr, lubridate
}
\examples{

}
