% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_Table_Var_Summary.R
\name{PM_Table_Var_Summary}
\alias{PM_Table_Var_Summary}
\title{Creates a Poshmark Summary Table by Variable Name}
\usage{
PM_Table_Var_Summary(pm_sales, pm_var)
}
\arguments{
\item{pm_sales}{The Poshmark Sales Table for the user.}

\item{pm_var}{The variable name for grouping}
}
\value{
A Poshmark Summary Table as a data frame.
}
\description{
This function Creates a Poshmark Summary Table by Variable Name.
}
\examples{

}
